Move points property to PointSet.
